Add protocol and extensions for warnings
To unify simple and detailed warnings
